Validate membership of a namespace
Rotter repos use namespaces to handle permissions; specifically whether a user is in the correct namespace or not.

This AuthoriseOperationFunc calls out to our graphql API to determine whether a particular user is a member of the wrapping namespace for the repo they want to clone.

If they are, then git operations succeed! Wonderful!

If not, then operations fail- as they should!